backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
aiosqlite>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
cryptography>=41.0.0
python-dotenv>=1.0.0
alembic>=1.12.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
aiofiles>=23.2.1
aiohttp>=3.9.0
slowapi>=0.1.8
python-multipart>=0.0.6
uvloop>=0.19.0; sys_platform != 'win32'
